requirements.txt
flask>=3.0
pyjwt>=2.8
cryptography>=42.0
jwcrypto>=1.5
pyyaml>=6.0
